Export Lock type
I have a slightly odd use case where I wish to obtain a lock,
write the lock data to disk, disconnect, then reload that lock data to
release the lock.

Without the Lock type being exported this makes it tricky.

Unfortunately due to a clash in naming I've had to rename a function,
but thought this was preferable to renaming the parameters in the Lock
struct.